+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [0.2.8] - 2025-01-22
+
+
+- Enhancements to `cloudly.gcp.bigquery`.
+
+
 ## [0.2.7] - 2025-01-15
 
 - New module `cloudly.gcp.bigquery`.